extract production code
