+++ lib/core/navigator/app_navigator_factory.dart
@@ -1,0 +1,10 @@
+import 'package:get/get.dart';
+
+import 'navigator_factory.dart';
+
+class AppNavigatorFactory extends NavigatorFactory {
+  @override
+  void offAndToNamed(String routeName) {
+    Get.offAndToNamed(routeName);
+  }
+}
