+++ lib/core/navigator/navigator_factory.dart
@@ -1,0 +1,3 @@
+abstract class NavigatorFactory {
+  void offAndToNamed(String routeName);
+}
